Extract fallback selection strategy
