+++ spec/lib/subjects/postgresql_selection_spec.rb
@@ -131,40 +131,5 @@ RSpec.describe Subjects::PostgresqlSelection do
         end
       end
     end
-
-    describe "#any_workflow_data" do
-      let(:subject_set_id) { nil }
-      let(:opts) { { limit: 5, subject_set_id: subject_set_id } }
-      let(:expected_ids) do
-        workflow.set_member_subjects.pluck("set_member_subjects.id")
-      end
-      let(:subject_ids) { subject.any_workflow_data }
-
-      it "should select some data from the workflow" do
-        expect(expected_ids).to include(*subject_ids)
-      end
-
-      context "grouped workflow" do
-        let(:subject_set_id) { SubjectSet.first.id }
-
-        before do
-          allow_any_instance_of(Workflow).to receive(:grouped).and_return(true)
-        end
-
-        it "should select some data from the group" do
-          expect(expected_ids).to include(*subject_ids)
-        end
-
-        context "without a subject_set_id param" do
-          let(:subject_set_id) { nil }
-
-          it "should raise an error" do
-            expect {
-              subject_ids
-            }.to raise_error(Subjects::Selector::MissingParameter)
-          end
-        end
-      end
-    end
   end
 end
